fix a missing comma in auto_update.json
This was spitting out a json parsing error whenever I launch SMM on windows 7.  Pretty simple syntax error, but it might be breaking... something.  I was having a problem where anytime I'd patch the game (even with no mods) it would corrupt the game so that it crashes immediately after loading.  After I fixed this file in my install, it seemed to work.